moved prompt processing into skills module
