+++ src/skills.rs
@@ -68,6 +68,69 @@ pub fn load_skill_content(path: &Path) -> Result<String> {
         .with_context(|| format!("Failed to read skill file: {}", path.display()))
 }
 
+// parses a prompt in case it contains a direct skill invocation
+pub fn process_prompt(input: String, skills: &SkillMapping) -> String {
+    let prompt = input.trim();
+    if let Some(skill_name_end) = prompt.find(' ').or_else(|| {
+        if prompt.starts_with('/') {
+            Some(prompt.len())
+        } else {
+            None
+        }
+    }) {
+        if prompt.starts_with('/') {
+            let skill_name = &prompt[1..skill_name_end];
+            if let Some(skill_path) = skills.get(skill_name) {
+                let remaining_prompt = if skill_name_end < prompt.len() {
+                    prompt[skill_name_end..].trim()
+                } else {
+                    ""
+                };
+
+                match std::fs::read_to_string(skill_path) {
+                    Ok(skill_content) => {
+                        if remaining_prompt.is_empty() {
+                            skill_content
+                        } else {
+                            format!("{}\n\n{}", skill_content, remaining_prompt)
+                        }
+                    }
+                    Err(e) => {
+                        eprintln!(
+                            "Warning: failed to read skill file '{}': {}",
+                            skill_path.display(),
+                            e
+                        );
+                        input
+                    }
+                }
+            } else {
+                input
+            }
+        } else {
+            input
+        }
+    } else if let Some(skill_name) = prompt.strip_prefix('/') {
+        if let Some(skill_path) = skills.get(skill_name) {
+            match std::fs::read_to_string(skill_path) {
+                Ok(skill_content) => skill_content,
+                Err(e) => {
+                    eprintln!(
+                        "Warning: failed to read skill file '{}': {}",
+                        skill_path.display(),
+                        e
+                    );
+                    input
+                }
+            }
+        } else {
+            input
+        }
+    } else {
+        input
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
